split integration test files
